single fetch method (single, array or variadic params)
