+++ spec/unit/repository/repository-fetch.spec.ts
@@ -96,5 +96,69 @@ describe("Repository", () => {
       it("returns the expected entity", () =>
         expect(entity).toEqual(expect.objectContaining({ entityId, ...data.expectedData })));
     });
+
+    const fullHashMock = { aString: A_STRING, aNumber: A_NUMBER_STRING, aBoolean: '0', someText: SOME_TEXT, aPoint: A_POINT_STRING, aDate: A_DATE_EPOCH.toString(), someStrings: SOME_STRINGS_JOINED };
+    const partialHashMock = { aString: A_STRING, aNumber: A_NUMBER_STRING };
+    const emptyHashMock = {}
+
+    const fullJsonMock = { aString: A_STRING, aNumber: A_NUMBER, aBoolean: false, someText: SOME_TEXT, aPoint: A_POINT_STRING, aDate: A_DATE_EPOCH, someStrings: SOME_STRINGS };
+    const partialJsonMock = { aString: A_STRING, aNumber: A_NUMBER };
+    const emptyJsonMock = {};
+
+    const expectedFull = { aString: A_STRING, aNumber: A_NUMBER, aBoolean: false, someText: SOME_TEXT, aPoint: A_POINT, aDate: A_DATE, someStrings: SOME_STRINGS };
+    const expectedPartial = { aString: A_STRING, aNumber: A_NUMBER, aBoolean: null, someText: null, aDate: null, aPoint: null, someStrings: null };
+    const expectedEmpty = { aString: null, aNumber: null, aBoolean: null, aDate: null, someText: null, aPoint: null, someStrings: null };
+
+    describe("when fetching a Hash", () => {
+
+      let repository: Repository<SimpleHashEntity>;
+      let entities: SimpleHashEntity[];
+
+      beforeAll(() => {
+        client = new Client()
+      });
+
+      beforeEach(async () => {
+        repository = new HashRepository(simpleHashSchema, client);
+        vi.mocked(client.hgetall)
+          .mockResolvedValueOnce(fullHashMock)
+          .mockResolvedValueOnce(partialHashMock)
+          .mockResolvedValue(emptyHashMock);
+        entities = await repository.fetch('foo', 'bar', 'baz');
+      });
+
+      it("returns the expected number of entities", () =>
+        expect(entities).toHaveLength(3));
+
+      it("returns the expected entities", () =>
+        expect(entities).toEqual(expect.arrayContaining([
+          expect.objectContaining({ entityId: 'foo', ...expectedFull }),
+          expect.objectContaining({ entityId: 'bar', ...expectedPartial }),
+          expect.objectContaining({ entityId: 'baz', ...expectedEmpty })
+        ])));
+    });
+
+    describe("when fetching JSON", () => {
+
+      let repository: Repository<SimpleJsonEntity>;
+      let entities: SimpleJsonEntity[];
+
+      beforeEach(async () => {
+        repository = new JsonRepository(simpleJsonSchema, client);
+        vi.mocked(client.jsonget)
+          .mockResolvedValueOnce(fullJsonMock)
+          .mockResolvedValueOnce(partialJsonMock)
+          .mockResolvedValue(emptyJsonMock)
+
+        entities = await repository.fetch('foo', 'bar', 'baz');
+      });
+
+      it("returns the expected entity", () =>
+        expect(entities).toEqual(expect.arrayContaining([
+          expect.objectContaining({ entityId: 'foo', ...expectedFull }),
+          expect.objectContaining({ entityId: 'bar', ...expectedPartial }),
+          expect.objectContaining({ entityId: 'baz', ...expectedEmpty })
+        ])));
+    });
   });
 });
